requirements.txt
Flask==3.1.0
flask_login==0.6.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
langchain_core==0.3.34
langchain_deepseek==0.1.1
langchain_openai==0.3.4
python-dotenv==1.0.1
qrcode==8.0
Werkzeug==3.1.3
gunicorn